requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
pydantic
langchain
langgraph
openai
discord
python-dotenv
requests